+++ src/balance.rs
@@ -5,11 +5,10 @@ use surrealdb::RecordId;
 use teloxide::types::ChatId;
 use travel_rs_derive::Table;
 
-const BALANCES_VIEW: &str = "v_balances";
+const FN_GET_BALANCES: &str = "fn::get_balances";
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Table)]
 pub struct Balance {
-    pub id: RecordId,
     pub debt: Decimal,
     pub debtor_name: Name,
     pub creditor_name: Name,
@@ -17,38 +16,40 @@ pub struct Balance {
 }
 
 impl Balance {
-    pub async fn db_select(chat_id: ChatId) -> Result<Vec<Self>, surrealdb::Error> {
-        use crate::chat::{ID as CHAT_ID, TABLE as CHAT_TB};
+    pub async fn balances(chat_id: ChatId) -> Result<Vec<Self>, surrealdb::Error> {
+        use crate::chat::TABLE as CHAT_TB;
 
         let db = db().await;
         db.query(format!(
             "SELECT *
-            FROM {BALANCES_VIEW}
-            WHERE {CHAT} = ${CHAT_ID}",
+            FROM {FN_GET_BALANCES}(${CHAT})
+            ORDER BY 
+                {DEBT} DESC, 
+                {DEBTOR_NAME} ASC, 
+                {CREDITOR_NAME} ASC",
         ))
-        .bind((CHAT_ID, RecordId::from_table_key(CHAT_TB, chat_id.0)))
+        .bind((CHAT, RecordId::from_table_key(CHAT_TB, chat_id.0)))
         .await
         .and_then(|mut response| response.take::<Vec<Self>>(0))
     }
 
-    pub async fn db_select_by_name(
+    pub async fn balances_by_name(
         chat_id: ChatId,
         name: Name,
     ) -> Result<Vec<Self>, surrealdb::Error> {
-        use crate::{
-            chat::{ID as CHAT_ID, TABLE as CHAT_TB},
-            traveler::NAME,
-        };
+        use crate::{chat::TABLE as CHAT_TB, traveler::NAME};
 
         let db = db().await;
         db.query(format!(
             "SELECT *
-            FROM {BALANCES_VIEW}
-            WHERE
-                {CHAT} = ${CHAT_ID}
-                && ({DEBTOR_NAME} = ${NAME} || {CREDITOR_NAME} = ${NAME})",
+            FROM {FN_GET_BALANCES}(${CHAT})
+            WHERE {DEBTOR_NAME} = ${NAME} || {CREDITOR_NAME} = ${NAME}
+            ORDER BY 
+                {DEBT} DESC, 
+                {DEBTOR_NAME} ASC, 
+                {CREDITOR_NAME} ASC",
         ))
-        .bind((CHAT_ID, RecordId::from_table_key(CHAT_TB, chat_id.0)))
+        .bind((CHAT, RecordId::from_table_key(CHAT_TB, chat_id.0)))
         .bind((NAME, name))
         .await
         .and_then(|mut response| response.take::<Vec<Self>>(0))