+++ src/tables/chat.rs
@@ -10,6 +10,7 @@ pub struct Chat {
     pub id: RecordId,
     pub last_interaction_utc: Datetime,
     pub lang: String,
+    pub currency: String,
 }
 
 impl Chat {
@@ -18,18 +19,24 @@ impl Chat {
         db.select((TABLE, id.0)).await
     }
 
-    pub async fn db_create(id: ChatId, lang: String) -> Result<Option<Self>, surrealdb::Error> {
+    pub async fn db_create(
+        id: ChatId,
+        langid: &LanguageIdentifier,
+        currency: &str,
+    ) -> Result<Option<Self>, surrealdb::Error> {
         let db = db().await;
         db.query(format!(
             "CREATE ${ID}
             CONTENT {{
                 {LAST_INTERACTION_UTC}: ${LAST_INTERACTION_UTC}, 
                 {LANG}: ${LANG},
+                {CURRENCY}: ${CURRENCY},
             }}",
         ))
         .bind((ID, RecordId::from_table_key(TABLE, id.0)))
         .bind((LAST_INTERACTION_UTC, Datetime::default()))
-        .bind((LANG, lang))
+        .bind((LANG, langid.to_string()))
+        .bind((CURRENCY, currency.to_owned()))
         .await
         .and_then(|mut response| response.take::<Option<Self>>(0))
     }
@@ -59,6 +66,21 @@ impl Chat {
         ))
         .bind((ID, RecordId::from_table_key(TABLE, id.0)))
         .bind((LANG, langid.to_string()))
+        .await
+        .and_then(|mut response| response.take::<Option<Self>>(0))
+    }
+
+    pub async fn db_update_currency(
+        id: ChatId,
+        currency: &str,
+    ) -> Result<Option<Self>, surrealdb::Error> {
+        let db = db().await;
+        db.query(format!(
+            "UPDATE ${ID}
+            SET {CURRENCY} = ${CURRENCY}",
+        ))
+        .bind((ID, RecordId::from_table_key(TABLE, id.0)))
+        .bind((CURRENCY, currency.to_owned()))
         .await
         .and_then(|mut response| response.take::<Option<Self>>(0))
     }